Added Comment interface, renamed Comment to CommentDTO, moved Entity to domain.
