+++ src/main/java/com/danigu/blog/comment/Comment.java
@@ -1,16 +1,12 @@
 package com.danigu.blog.comment;
 
 import com.danigu.blog.post.Post;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
 
 /**
  * @author dani
  */
-@AllArgsConstructor
-@Getter
-public class Comment {
-    public long id;
-    public Post post;
-    public String content;
+public interface Comment {
+    long getId();
+    String getContent();
+    Post getPost();
 }
