+++ src/main/java/com/danigu/blog/comment/CommentEntity.java
@@ -1,4 +1,4 @@
-package com.danigu.blog.comment.persistence;
+package com.danigu.blog.comment;
 
 import com.danigu.blog.post.persistence.PostEntity;
 import lombok.AllArgsConstructor;
@@ -9,9 +9,10 @@ import lombok.Value;
 import javax.persistence.*;
 
 /**
+ * The mapping could've been done with XML for better separation, although this is not tied to the implementation.
  * @author dani
  */
-@Entity(name = "Comment")
+@Entity(name = "CommentDto")
 @Table(name = "comment")
 @Data
 @NoArgsConstructor
