+++ src/main/resources/META-INF/persistence.xml
@@ -8,7 +8,7 @@
     <persistence-unit name="blog" transaction-type="RESOURCE_LOCAL">
         <provider>org.hibernate.jpa.HibernatePersistenceProvider</provider>
         <class>com.danigu.blog.post.persistence.PostEntity</class>
-        <class>com.danigu.blog.comment.persistence.CommentEntity</class>
+        <class>com.danigu.blog.commentDto.CommentEntity</class>
         <properties>
             <property name="hibernate.connection.url" value="jdbc:mysql://localhost:3306/blog"/>
             <property name="hibernate.connection.username" value="root"/>
